vit-image-classifier/requirements (1).txt
# For local inference server (server.py)
# Not needed if using the HuggingFace Inference API (default)

torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
